requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
git+https://github.com/the-strategy-unit/nhp_aci.git#egg=nhp_aci
